move the gunk in testing.__init__ to testing.subprocess
