+++ tests/spec/configsearch.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 import uuid
 from sys import executable
 
-from testing import assert_command
+from testing.subprocess import assert_command
 
 
 class DescribeCombined(object):
